requirements.txt
#TA_Lib==0.4.19
numpy==1.19.5
arch==4.19
matplotlib==3.3.4
xgboost==1.4.1
pandas==1.2.3
mlxtend==0.18.0
yfinance==0.1.59
streamlit==0.80.0
bt==0.2.9
plotly==4.14.3
python_dateutil==2.8.1
scikit_learn==0.24.2
